fix: logical error in inner_insert_ and inner_insert_at_
If the current node is split, but key should still be inserted into this
node, it passes down to the insertion using the passed-in index (which
may very well be obsolete).

This fixes that by calling inner_insert_ on this node when the key shall
still be inserted into this node, and its new neighbor otherwise. After
that, the function ends (meaning the obsolete `index` is never used).

commit-by: Huy Nguyen