doc: update comments and documentations
Inside inner_insert_, add comment about transferring ownership of the
child larger than the node's median into the 0th position of the new
node.

Inside inner_insert_, the branch where `this` is root node, comment that
curr_bt->root_ is an unique_ptr to `this` (which explains the
std::move()). FIX: this comment may be redundant.

In inner_insert_child_at_ documentation, update conditions (and
subsequently, the corresponding assert()'s inside the function).

In BTree's insert(), specify the behaviour when the function returns
false.