feat: finish declaring remove functions
Also, change nonleaf_remove[_at] to take the smallest of right subtree.
This is because the [non]leaf_rebalance_ method defaults to merging with
the left node if available. If kept as taking largest of left subtree,
the left node will be merged with the left of left, which is not what we
want.
Now it's defining time.

commit-by: Huy Nguyen