i'm so done with this
Turns out the fuzz test breaks not because of the BTree being faulty,
but because the fuzz test itself being faulty.

The unit test works very well indeed.

Very thanks to AddressSanitizer that I caught that pesky little sh-

Some further changes may be made to the CMake scripts to better
interoperate with the main project.

commit-by: Huy Nguyen